requirements-streamlit.txt
streamlit==1.56.0
plotly==6.7.0
pandas==3.0.2
psycopg2-binary==2.9.12
